@@ -3,72 +3,74 @@
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
-    "ng": "nx",
-    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2020 browser module main",
     "start": "nx serve",
     "build": "nx build",
     "test": "nx test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "14.2.4",
-    "@angular/cdk": "14.2.3",
-    "@angular/common": "14.2.4",
-    "@angular/compiler": "14.2.4",
-    "@angular/core": "14.2.4",
-    "@angular/forms": "14.2.4",
-    "@angular/material": "14.2.3",
-    "@angular/platform-browser": "14.2.4",
-    "@angular/platform-browser-dynamic": "14.2.4",
-    "@angular/router": "14.2.4",
-    "@nestjs/common": "9.1.2",
-    "@nestjs/core": "9.1.2",
-    "@nestjs/platform-express": "9.1.2",
-    "@ngrx/component-store": "14.0.2",
-    "@ngrx/effects": "14.0.2",
-    "@ngrx/entity": "14.0.2",
-    "@ngrx/router-store": "14.0.2",
-    "@ngrx/store": "14.0.2",
-    "@ngrx/store-devtools": "14.0.2",
-    "@nrwl/angular": "14.8.2",
+    "@angular/animations": "17.0.0-rc.2",
+    "@angular/cdk": "17.0.0-rc.2",
+    "@angular/common": "17.0.0-rc.2",
+    "@angular/compiler": "17.0.0-rc.2",
+    "@angular/core": "17.0.0-rc.2",
+    "@angular/forms": "17.0.0-rc.2",
+    "@angular/material": "17.0.0-rc.2",
+    "@angular/platform-browser": "17.0.0-rc.2",
+    "@angular/platform-browser-dynamic": "17.0.0-rc.2",
+    "@angular/router": "17.0.0-rc.2",
+    "@nestjs/common": "10.2.8",
+    "@nestjs/core": "10.2.8",
+    "@nestjs/platform-express": "10.2.8",
+    "@ngrx/component-store": "17.0.0-beta.0",
+    "@ngrx/effects": "17.0.0-beta.0",
+    "@ngrx/entity": "17.0.0-beta.0",
+    "@ngrx/operators": "^17.0.0-beta.0",
+    "@ngrx/router-store": "17.0.0-beta.0",
+    "@ngrx/signals": "^17.0.0-beta.0",
+    "@ngrx/store": "17.0.0-beta.0",
+    "@ngrx/store-devtools": "17.0.0-beta.0",
+    "@nx/angular": "17.1.0-beta.2",
     "reflect-metadata": "^0.1.13",
-    "rxjs": "~7.5.0",
+    "rxjs": "7.8.1",
     "tslib": "^2.0.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "^0.14.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "14.2.4",
-    "@angular-eslint/eslint-plugin": "14.0.4",
-    "@angular-eslint/eslint-plugin-template": "14.0.4",
-    "@angular-eslint/template-parser": "14.0.4",
-    "@angular/cli": "~14.2.0",
-    "@angular/compiler-cli": "14.2.4",
-    "@angular/language-service": "14.2.4",
-    "@nestjs/schematics": "9.0.3",
-    "@nestjs/testing": "9.1.2",
-    "@nrwl/cli": "14.8.2",
-    "@nrwl/cypress": "14.8.2",
-    "@nrwl/eslint-plugin-nx": "14.8.2",
-    "@nrwl/jest": "14.8.2",
-    "@nrwl/linter": "14.8.2",
-    "@nrwl/nest": "14.8.2",
-    "@nrwl/node": "14.8.2",
-    "@nrwl/workspace": "14.8.2",
-    "@types/jest": "28.1.8",
+    "@angular-devkit/build-angular": "17.0.0-rc.2",
+    "@angular-devkit/core": "17.0.0-rc.2",
+    "@angular-devkit/schematics": "17.0.0-rc.2",
+    "@angular-eslint/eslint-plugin": "16.0.3",
+    "@angular-eslint/eslint-plugin-template": "16.0.3",
+    "@angular-eslint/template-parser": "16.0.3",
+    "@angular/cli": "17.0.0-rc.2",
+    "@angular/compiler-cli": "17.0.0-rc.2",
+    "@angular/language-service": "17.0.0-rc.2",
+    "@nestjs/schematics": "10.0.3",
+    "@nestjs/testing": "10.2.8",
+    "@nx/cypress": "17.1.0-beta.2",
+    "@nx/eslint": "17.1.0-beta.2",
+    "@nx/eslint-plugin": "17.1.0-beta.2",
+    "@nx/jest": "17.1.0-beta.2",
+    "@nx/nest": "17.1.0-beta.2",
+    "@nx/node": "17.1.0-beta.2",
+    "@nx/workspace": "17.1.0-beta.2",
+    "@schematics/angular": "16.2.9",
+    "@types/jest": "29.4.4",
     "@types/node": "18.7.1",
-    "@typescript-eslint/eslint-plugin": "5.38.1",
-    "@typescript-eslint/parser": "5.38.1",
+    "@typescript-eslint/eslint-plugin": "6.9.1",
+    "@typescript-eslint/parser": "6.9.1",
     "cypress": "^9.1.0",
-    "eslint": "8.15.0",
-    "eslint-config-prettier": "8.1.0",
-    "eslint-plugin-cypress": "^2.10.3",
-    "jest": "28.1.3",
-    "jest-environment-jsdom": "28.1.1",
-    "jest-preset-angular": "12.2.2",
+    "eslint": "8.46.0",
+    "eslint-config-prettier": "9.0.0",
+    "eslint-plugin-cypress": "2.15.1",
+    "jest": "29.4.3",
+    "jest-environment-jsdom": "29.4.3",
+    "jest-preset-angular": "13.1.2",
+    "nx": "17.1.0-beta.2",
     "prettier": "2.7.1",
-    "ts-jest": "28.0.8",
+    "ts-jest": "29.1.1",
     "ts-node": "10.9.1",
-    "typescript": "4.8.4",
-    "nx": "14.8.2"
+    "typescript": "5.2"
   }
 }
