@@ -1,41 +1,34 @@
 import { Component } from '@angular/core';
-import { combineLatest, Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { from, map, mergeMap, Observable, switchMap, toArray } from 'rxjs';
 
-import { CartProduct } from '../../model/product';
-import { CartService } from '../cart.service';
-import { ProductService } from '../../product/product.service';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { Router } from '@angular/router';
+import { CartProduct } from '../../model/product';
+import { ProductService } from '../../product/product.service';
+import { CartService } from '../cart.service';
 
 @Component({
   selector: 'ngrx-nx-workshop-cart-details',
   templateUrl: './cart-details.component.html',
-  styleUrls: ['./cart-details.component.scss']
+  styleUrls: ['./cart-details.component.scss'],
 })
 export class CartDetailsComponent {
-  cartProducts$: Observable<CartProduct[] | undefined> = combineLatest(
-    this.cartService.cartItems$,
-    this.productService.getProducts()
-  ).pipe(
-    map(([cartItems, products]) => {
-      if (!cartItems || !products) return undefined;
-      return cartItems
-        .map(({ productId, quantity }): CartProduct | undefined => {
-          const product = products.find(p => p.id === productId);
-          if (!product) return undefined;
-          return {
-            ...product,
-            quantity
-          };
-        })
-        .filter((cartProduct): cartProduct is CartProduct => !!cartProduct);
-    })
+  cartProducts$: Observable<CartProduct[]> = this.cartService.cartItems$.pipe(
+    switchMap((cartItems) =>
+      from(cartItems).pipe(
+        mergeMap((item) =>
+          this.productService
+            .getProduct(item.productId)
+            .pipe(map((product) => ({ ...product, quantity: item.quantity })))
+        ),
+        toArray()
+      )
+    )
   );
 
   total$ = this.cartProducts$.pipe(
     map(
-      cartProducts =>
+      (cartProducts) =>
         cartProducts &&
         cartProducts.reduce(
           (acc, product) => acc + product.price * product.quantity,
@@ -67,13 +60,13 @@ export class CartDetailsComponent {
         products.map(({ id, quantity }) => ({ productId: id, quantity }))
       )
       // 👇 really important not to forget to subscribe
-      .subscribe(isSuccess => {
+      .subscribe((isSuccess) => {
         if (isSuccess) {
           this.cartService.getCartProducts();
           this.router.navigateByUrl('');
         } else {
           this.snackBar.open('Purchase error', 'Error', {
-            duration: 2500
+            duration: 2500,
           });
         }
       });
