@@ -2,8 +2,16 @@ import { Location } from '@angular/common';
 import { Component } from '@angular/core';
 import { ActivatedRoute, ParamMap } from '@angular/router';
 import { Rating } from '@ngrx-nx-workshop/api-interfaces';
-import { BehaviorSubject, Observable, of } from 'rxjs';
-import { filter, map, shareReplay, switchMap } from 'rxjs/operators';
+import {
+  BehaviorSubject,
+  combineLatest,
+  concatMap,
+  filter,
+  map,
+  shareReplay,
+  switchMap,
+  take,
+} from 'rxjs';
 
 import { CartService } from '../../cart/cart.service';
 import { ProductService } from '../product.service';
@@ -12,7 +20,7 @@ import { RatingService } from '../rating.service';
 @Component({
   selector: 'ngrx-nx-workshop-product-details',
   templateUrl: './product-details.component.html',
-  styleUrls: ['./product-details.component.scss']
+  styleUrls: ['./product-details.component.scss'],
 })
 export class ProductDetailsComponent {
   private readonly productId$ = this.router.paramMap.pipe(
@@ -21,11 +29,20 @@ export class ProductDetailsComponent {
     shareReplay({ bufferSize: 1, refCount: true })
   );
 
-  product$ = this.productId$.pipe(
-    switchMap(id => this.productService.getProduct(id))
+  readonly product$ = this.productId$.pipe(
+    switchMap((id) => this.productService.getProduct(id))
   );
 
-  private customerRating$ = new BehaviorSubject<number | undefined>(undefined);
+  readonly reviewsRefresh$ = new BehaviorSubject<void>(undefined);
+
+  readonly reviews$ = combineLatest([
+    this.productId$,
+    this.reviewsRefresh$,
+  ]).pipe(switchMap(([id]) => this.ratingService.getReviews(id)));
+
+  protected customerRating$ = new BehaviorSubject<number | undefined>(
+    undefined
+  );
 
   constructor(
     private readonly router: ActivatedRoute,
@@ -35,8 +52,8 @@ export class ProductDetailsComponent {
     private readonly location: Location
   ) {
     this.productId$
-      .pipe(switchMap(id => this.ratingService.getRating(id)))
-      .subscribe(productRating =>
+      .pipe(switchMap((id) => this.ratingService.getRating(id)))
+      .subscribe((productRating) =>
         this.customerRating$.next(productRating && productRating.rating)
       );
   }
@@ -45,16 +62,16 @@ export class ProductDetailsComponent {
     this.ratingService
       .setRating({ productId, rating })
       .pipe(
-        map(arr =>
-          arr.find(productRating => productId === productRating.productId)
+        map((arr) =>
+          arr.find((productRating) => productId === productRating.productId)
         ),
         filter(
           (productRating): productRating is NonNullable<typeof productRating> =>
             productRating != null
         ),
-        map(productRating => productRating.rating)
+        map((productRating) => productRating.rating)
       )
-      .subscribe(newRating => this.customerRating$.next(newRating));
+      .subscribe((newRating) => this.customerRating$.next(newRating));
   }
 
   addToCart(productId: string) {
@@ -63,5 +80,21 @@ export class ProductDetailsComponent {
 
   back() {
     this.location.back();
+  }
+
+  submit(review: { reviewer: string; reviewText: string }) {
+    this.productId$
+      .pipe(
+        take(1),
+        concatMap((productId) =>
+          this.ratingService.postReview({
+            productId,
+            ...review,
+          })
+        )
+      )
+      .subscribe(() => {
+        this.reviewsRefresh$.next();
+      });
   }
 }
