@@ -9,7 +9,7 @@ describe('AppController', () => {
   beforeAll(async () => {
     app = await Test.createTestingModule({
       controllers: [AppController],
-      providers: [ProductService]
+      providers: [ProductService],
     }).compile();
   });
 
