@@ -4,13 +4,14 @@ import {
   Get,
   Param,
   Post,
-  UseInterceptors
+  UseInterceptors,
 } from '@nestjs/common';
 import {
   BasicProduct,
+  CartItem,
   Product,
   ProductRating,
-  CartItem
+  Review,
 } from '@ngrx-nx-workshop/api-interfaces';
 
 import { CartService } from './cart/cart.service';
@@ -75,5 +76,17 @@ export class AppController {
   @Get('rating/get-ratings')
   getRatings(): ProductRating[] {
     return this.ratingService.getRatings();
+  }
+
+  @Get('reviews/get/:productId')
+  getReviews(@Param('productId') productId: string): Review[] {
+    return this.ratingService.getReviews(productId);
+  }
+
+  @Post('reviews/post/:id')
+  addReview(
+    @Body() review: Pick<Review, 'productId' | 'reviewText' | 'reviewer'>
+  ): Review {
+    return this.ratingService.addReview(review);
   }
 }
