@@ -1,9 +1,11 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
-import { Product, BasicProduct } from '@ngrx-nx-workshop/api-interfaces';
+import { BasicProduct, Product } from '@ngrx-nx-workshop/api-interfaces';
 import { data } from '@ngrx-nx-workshop/data';
 
 function stripDescription(originalData: Product[]): BasicProduct[] {
-  return originalData.map(({description, ...product}) => product);
+  // Remove `description` from the object.
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  return originalData.map(({ description, ...product }) => product);
 }
 
 @Injectable()
@@ -13,7 +15,7 @@ export class ProductService {
   }
 
   getProduct(id: string): Product {
-    const product = data.find(p => p.id === id);
+    const product = data.find((p) => p.id === id);
     if (!product) {
       throw new NotFoundException(`Product with id ${id} is not found`);
     }
