@@ -3,6 +3,6 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'ngrx-nx-workshop-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss']
+  styleUrls: ['./app.component.scss'],
 })
 export class AppComponent {}
