@@ -1,11 +1,11 @@
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NgModule } from '@angular/core';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
-import { AppComponent } from './app.component';
-import { RoutingModule } from './router/routing.module';
-import { CartIconModule } from './cart/cart-icon/cart-icon.module';
-import { MatToolbarModule } from '@angular/material/toolbar';
 import { HttpClientModule } from '@angular/common/http';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { AppComponent } from './app.component';
+import { CartIconModule } from './cart/cart-icon/cart-icon.module';
+import { RoutingModule } from './router/routing.module';
 
 @NgModule({
   declarations: [AppComponent],
@@ -14,8 +14,8 @@ import { HttpClientModule } from '@angular/common/http';
     HttpClientModule,
     RoutingModule,
     CartIconModule,
-    MatToolbarModule
+    MatToolbarModule,
   ],
-  bootstrap: [AppComponent]
+  bootstrap: [AppComponent],
 })
 export class AppModule {}
