@@ -3,9 +3,9 @@ import { NgModule } from '@angular/core';
 import { MatCardModule } from '@angular/material/card';
 import { RouterModule } from '@angular/router';
 
-import { ProductListComponent } from './product-list.component';
-import { StarsModule } from '../../common/stars/stars.module';
 import { SpinnerModule } from '../../common/spinner/spinner.module';
+import { StarsModule } from '../../common/stars/stars.module';
+import { ProductListComponent } from './product-list.component';
 
 @NgModule({
   imports: [
@@ -13,9 +13,9 @@ import { SpinnerModule } from '../../common/spinner/spinner.module';
     MatCardModule,
     StarsModule,
     RouterModule,
-    SpinnerModule
+    SpinnerModule,
   ],
   declarations: [ProductListComponent],
-  exports: [ProductListComponent]
+  exports: [ProductListComponent],
 })
 export class ProductListModule {}
