@@ -1,6 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { ProductRating } from '@ngrx-nx-workshop/api-interfaces';
+import { ProductRating, Review } from '@ngrx-nx-workshop/api-interfaces';
 import { Observable } from 'rxjs';
 
 @Injectable({ providedIn: 'root' })
@@ -18,7 +18,22 @@ export class RatingService {
     return this.http.get<ProductRating[]>('/api/rating/get-ratings');
   }
 
-  getRating(id: string): Observable<ProductRating | undefined> {
-    return this.http.get<ProductRating | undefined>(`/api/rating/get/${id}`);
+  getRating(productId: string): Observable<ProductRating | undefined> {
+    return this.http.get<ProductRating | undefined>(
+      `/api/rating/get/${productId}`
+    );
+  }
+
+  getReviews(productId: string): Observable<Review[]> {
+    return this.http.get<Review[]>(`/api/reviews/get/${productId}`);
+  }
+
+  postReview(
+    review: Pick<Review, 'productId' | 'reviewText' | 'reviewer'>
+  ): Observable<Review> {
+    return this.http.post<Review>(
+      `/api/reviews/post/${review.productId}`,
+      review
+    );
   }
 }
