@@ -2,16 +2,16 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
 import { CartDetailsComponent } from '../cart/cart-details/cart-details.component';
-import { ProductListComponent } from '../product/product-list/product-list.component';
+import { CartDetailsModule } from '../cart/cart-details/cart-details.module';
 import { ProductDetailsComponent } from '../product/product-details/product-details.component';
 import { ProductDetailsModule } from '../product/product-details/product-details.module';
+import { ProductListComponent } from '../product/product-list/product-list.component';
 import { ProductListModule } from '../product/product-list/product-list.module';
-import { CartDetailsModule } from '../cart/cart-details/cart-details.module';
 
 const routes: Routes = [
   { path: 'details/:productId', component: ProductDetailsComponent },
   { path: 'cart', component: CartDetailsComponent },
-  { path: '', component: ProductListComponent, pathMatch: 'full' }
+  { path: '', component: ProductListComponent, pathMatch: 'full' },
 ];
 
 @NgModule({
@@ -19,8 +19,8 @@ const routes: Routes = [
     ProductDetailsModule,
     ProductListModule,
     CartDetailsModule,
-    RouterModule.forRoot(routes, { relativeLinkResolution: 'legacy' })
+    RouterModule.forRoot(routes),
   ],
-  exports: [RouterModule]
+  exports: [RouterModule],
 })
 export class RoutingModule {}
